refactor(issues): scope IssuesList back to /bounties star column
