+++ src/components/issues/index.ts
@@ -1,10 +1,7 @@
 /**
  * Issues components
  */
-export {
-  default as IssuesList,
-  ViewModeToggle as IssuesViewModeToggle,
-} from './IssuesList';
+export { default as IssuesList } from './IssuesList';
 export { type IssuesViewMode, ISSUES_VIEW_QUERY_PARAM } from './issuesViewMode';
 export { default as IssueStats } from './IssueStats';
 export { default as BountyProgress } from './BountyProgress';
